+++ open-bible-study/src/app/api/login/route.ts
@@ -1,4 +1,4 @@
-import { login } from "@/modules/auth/actions";
+import { login } from "@/features/auth/actions";
 
 export async function GET() {
   await login();
