+++ open-bible-study/src/features/study/components/bible-navigation/book-navigation.tsx
