+++ open-bible-study/src/features/study/components/bible-navigation/index.tsx
