+++ open-bible-study/src/features/study/components/bible-navigation/version-navigation.tsx
