+++ open-bible-study/src/features/study/components/bible/footer-navigation.tsx
