+++ open-bible-study/src/features/study/components/bible/index.tsx
