+++ open-bible-study/src/features/study/components/bible/no-verse-selected.tsx
