+++ open-bible-study/src/features/study/components/study-tools/index.tsx
