+++ open-bible-study/src/features/study/templates/study-page-template.tsx
